Fix error with using {.}
I was using version 0.1.6 and in my code, I use `{.|js|s}` to store the dust context. This fails in version 0.3.0 when my opt.context is an array of string JSON files. This line was removed for some reason, but it should be added back for people who use `{.}` in their code.